% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepforestr.R
\name{get_data}
\alias{get_data}
\title{Check for updates}
\usage{
get_data(repo = "")
}
\arguments{
\item{repo}{path to the repository}
}
\description{
Check for updates
}
\examples{
\donttest{
deepforest::get_data()
}
}
