% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepforestr.R
\name{install_deepforest}
\alias{install_deepforest}
\title{Install the DeepForest Python package}
\usage{
install_deepforest()
}
\description{
Install the DeepForest Python package
}
\examples{
\dontrun{
deepforestr::install_deepforest()}

}
