% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepforestr.R
\name{get_data}
\alias{get_data}
\title{Get example data}
\usage{
get_data(path)
}
\arguments{
\item{path}{path to the example data file}
}
\description{
Get example data
}
\examples{
\dontrun{
deepforestr::get_data("OSBS_029.png")
}

}
